+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -5,6 +5,7 @@
 This version and all future versions will require Python 3.7+. Python 3.6 is no longer supported.
 
 ### Features Added
+- Added support for `AzureNamedKeyCredential` as a valid `credential` type.
 
 ### Bugs Fixed